feat: Fix analysis issues
