+++ lib/src/functions_model/functions_model.dart
@@ -1,1 +1,2 @@
 export 'accept_quote.dart';
+export 'update_delivery.dart';
